group papers together in collapsible sub menu
makes it easier to collapse/open papers and find content
